Introduce concept of a "provider" initially for decoders & implemented for wavpack.
Instead of registering a factory with a manager, registering a provider
can give much better capabilities. This is not yet plumbed in.

A provider can be queried for support of a mime type, verify by some
means (MD5 exposed in decoder.hpp for convenience), and create a decoder
object.

This also makes for easier testing as noted by the expanded wavpack test
suite.